+++ pages/dev/src/collections/pages.ts
@@ -19,6 +19,7 @@ export const Pages: PageCollectionConfig = {
     },
     isRootCollection: true,
   },
+  trash: true,
   versions: {
     drafts: true,
   },