+++ pages/dev_multi_tenant/src/collections/pages.ts
@@ -16,6 +16,7 @@ export const Pages: PageCollectionConfig = {
       unique: false,
     },
   },
+  trash: true,
   versions: {
     drafts: true,
   },